platform: add and use nm_sock_addr_union_to_string()
We still don't use getnameinfo(). This is used for logging,
where we want to see the a string representation that is possibly
close to the actual bytes (to spot differences). It should not be
obfuscated by a function out of our control.

Also fix the notation for the IPv6 scope ID to use the common '%'
character.